refactor | Update imports and remove files does not used
