--- app/components/ui/index.ts
@@ -1,1 +1,0 @@
-export * from './horizontalBar/HorizontalBar';
