--- app/components/ui/map/MapContainer.tsx
@@ -1,34 +1,0 @@
-"use client"
-import React, {useEffect} from 'react'
-
-import {
-    APIProvider,
-    Map,
-    Marker,
-    useMarkerRef
-} from '@vis.gl/react-google-maps';
-
-export const MapContainer = () => {
-
-    const [markerRef, marker] = useMarkerRef();
-    
-    useEffect(() => {
-        if (!marker) {
-          return;
-        }
-    
-        // do something with marker instance here
-      }, [marker]);
-
-    return (
-        <APIProvider apiKey={`${process.env.NEXT_PUBLIC_API_KEY}`}>
-            <Map
-                defaultCenter={{ lat: 25.698425, lng: -100.34507 }}
-                defaultZoom={12}
-                className='w-full h-full'
-            >
-                <Marker ref={markerRef} position={{ lat: 25.698425, lng: -100.34507 }} />
-            </Map>
-        </APIProvider>
-    )
-}
