--- app/components/wrappers/index.ts
@@ -1,1 +1,0 @@
-export * from './cardWrapper/CardWrapper';
