+++ app/lib/constants.ts
@@ -1,0 +1,100 @@
+export const devicesColumns=[
+    {
+        id:'unit',
+        field:'device',
+        header: 'Unidad'
+    }
+]
+export const detailDeviceColumns = [
+    {
+        id: '1',
+        field: 'date',
+        header: 'Fecha'
+    },
+    {
+        id: '2',
+        field: 'codeSeal',
+        header: 'Código de sello'
+    },
+    {
+        id: '3',
+        field: 'numberCompartment',
+        header: 'Compartimiento'
+    },
+    {
+        id: '4',
+        field: 'motorStatus',
+        header: 'Estado del motor'
+    },
+    {
+        id: '5',
+        field: 'capacityDevice',
+        header: 'Capacidad de la unidad'
+    },
+    {
+        id: '6',
+        field: 'lastLocation',
+        header: 'Últimas coordenadas'
+    }
+];
+
+export const columns = [
+    {
+        id: '1',
+        field: 'name',
+        header: 'Unidad'
+    },
+    {
+        id: '2',
+        field: 'trailer',
+        header: 'Remolque'
+    },
+    {
+        id: '3',
+        field: 'date',
+        header: 'Fecha'
+    },
+    {
+        id: '4',
+        field: 'status',
+        header: 'Estatus unidad'
+    },
+    {
+        id: '5',
+        field: 'codeSeal',
+        header: 'Código de sello'
+    }
+];
+
+export const detailedColumns = [
+    {
+        id: '1',
+        field:'name',
+        header: 'Unidad'
+    },
+    {
+        id: '2',
+        field: 'group',
+        header: 'Grupo'
+    },
+    {
+        id: '3',
+        field: 'driver',
+        header: 'Conductor'
+    },
+    {
+        id: '4',
+        field: 'numberOfTrailers',
+        header: 'Remolque'
+    },
+    {
+        id: '5',
+        field: 'lastCodeSeal',
+        header: 'Último sello'
+    },
+    {
+        id: '6',
+        field: 'lastLocation',
+        header: 'Última ubicación'
+    }
+];
