--- app/lib/data.ts
@@ -1,427 +1,0 @@
-export const devices = [
-    {
-        id :'1',
-        name: 'TRC213',
-        group: 'Unidad Pipa de Gas 2',
-        driver: 'Sergio Pérez',
-        trailer : 'Remolque 1',
-        numberOfTrailers: 1,
-        date: '19/03/2023 | 06:35:23',
-        status: 'Ralentí',
-        codeSeal : '377852',
-        lastCodeSeal: '377852',
-        lastLocation: '25.752935, -100.5066483'
-    },
-    {
-        id :'2',
-        name: 'TRC214',
-        group: 'Grupo Utilitario',
-        driver: 'Sergio Pérez',
-        trailer : 'Remolque 2',
-        numberOfTrailers: 1,
-        date: '19/03/2023 | 06:35:23',
-        status: 'Ralentí',
-        codeSeal : '377853',
-        lastCodeSeal: '377853',
-        lastLocation: '25.752935, -100.5066483'
-    },
-    {
-        id :'3',
-        name: 'TRC215',
-        trailer : 'Remolque 6',
-        date: '19/03/2023 | 06:35:23',
-        status: 'En movimiento',
-        codeSeal : '270054',
-        group: 'Grupo Utilitario',
-        driver: 'Sergio Pérez',
-        numberOfTrailers: 1,
-        lastCodeSeal: '270054',
-        lastLocation: '25.752935, -100.5066483'
-    },
-    {
-        id :'4',
-        name: 'TRC213',
-        group: 'Grupo Utilitario',
-        driver: 'Sergio Pérez',
-        trailer : 'Remolque 8',
-        date: '19/03/2023 | 06:35:23',
-        status: 'Ralentí',
-        codeSeal : '270066',
-        numberOfTrailers: 1,
-        lastCodeSeal: '270066',
-        lastLocation: '25.752935, -100.5066483'
-    
-    },
-    {
-        id :'5',
-        name: 'TRC216',
-        trailer : 'Remolque 3',
-        date: '19/03/2023 | 06:35:23',
-        status: 'Ralentí',
-        codeSeal : '270067',
-        group: 'Grupo Utilitario',
-        driver: 'Sergio Pérez',
-        numberOfTrailers: 1,
-        lastCodeSeal: '270067',
-        lastLocation: '25.752935, -100.5066483'
-    
-    },
-    {
-        id :'6',
-        name: 'TRC217',
-        trailer : 'Remolque 28',
-        date: '19/03/2023 | 06:35:23',
-        status: 'Ralentí',
-        codeSeal : '270066',
-        group: 'Grupo Utilitario',
-        driver: 'Sergio Pérez',
-        numberOfTrailers: 2,
-        lastCodeSeal: '270066',
-        lastLocation: '25.752935, -100.5066483'
-    },
-    {
-        id :'7',
-        name: 'TRC218',
-        trailer : 'Remolque 16',
-        date: '19/03/2023 | 06:35:23',
-        status: 'Ralentí',
-        codeSeal : '270069',
-        group: 'Grupo Utilitario',
-        driver: 'Sergio Pérez',
-        numberOfTrailers: 1,
-        lastCodeSeal: '270069',
-        lastLocation: '25.752935, -100.5066483'
-    }, 
-    {
-        id :'8',
-        name: 'TRC219',
-        trailer : 'Remolque 17',
-        date: '19/03/2023 | 06:35:23',
-        status: 'En movimiento',
-        codeSeal : '270070',
-        group: 'Grupo Utilitario',
-        driver: 'Sergio Pérez',
-        numberOfTrailers: 1,
-        lastCodeSeal: '270070',
-        lastLocation: '25.752935, -100.5066483'
-    },
-    {
-        id :'9',
-        name: 'TRC2120',
-        trailer : 'Remolque 16',
-        date: '19/03/2023 | 06:35:23',
-        status: 'Ralentí',
-        codeSeal : '270071',
-        group: 'Grupo Utilitario',
-        driver: 'Sergio Pérez',
-        numberOfTrailers: 2,
-        lastCodeSeal: '270071',
-        lastLocation: '25.752935, -100.5066483'
-    }, 
-    {
-        id :'10',
-        name: 'TRC2121',
-        trailer : 'Remolque 17',
-        date: '19/03/2023 | 06:35:23',
-        status: 'En movimiento',
-        codeSeal : '270080',
-        group: 'Grupo Utilitario',
-        driver: 'Sergio Pérez',
-        numberOfTrailers: 2,
-        lastCodeSeal: '270080',
-        lastLocation: '25.752935, -100.5066483'
-    },
-    {
-        id :'11',
-        name: 'TRC222',
-        trailer : 'Remolque 16',
-        date: '19/03/2023 | 06:35:23',
-        status: 'Ralentí',
-        codeSeal : '270098',
-        group: 'Grupo Utilitario',
-        driver: 'Sergio Pérez',
-        numberOfTrailers: 2,
-        lastCodeSeal: '270098',
-        lastLocation: '25.752935, -100.5066483'
-    }, 
-    {
-        id :'12',
-        name: 'TRC223',
-        trailer : 'Remolque 17',
-        date: '19/03/2023 | 06:35:23',
-        status: 'En movimiento',
-        codeSeal : '270067',
-        group: 'Grupo Utilitario',
-        driver: 'Sergio Pérez',
-        numberOfTrailers: 2,
-        lastCodeSeal: '270067',
-        lastLocation: '25.752935, -100.5066483'
-    },
-    {
-        id :'13',
-        name: 'TRC224',
-        trailer : 'Remolque 16',
-        date: '19/03/2023 | 06:35:23',
-        status: 'Ralentí',
-        codeSeal : '270068',
-        group: 'Grupo Utilitario',
-        driver: 'Sergio Pérez',
-        numberOfTrailers: 2,
-        lastCodeSeal: '270068',
-        lastLocation: '25.752935, -100.5066483'
-    }, 
-    {
-        id :'14',
-        name: 'TRC225',
-        trailer : 'Remolque 17',
-        date: '19/03/2023 | 06:35:23',
-        status: 'En movimiento',
-        codeSeal : '270069',
-        group: 'Grupo Utilitario',
-        driver: 'Sergio Pérez',
-        numberOfTrailers: 2,
-        lastCodeSeal: '270069',
-        lastLocation: '25.752935, -100.5066483'
-    },
-    {
-        id :'15',
-        name: 'TRC226',
-        trailer : 'Remolque 16',
-        date: '19/03/2023 | 06:35:23',
-        status: 'Ralentí',
-        codeSeal : '270070',
-        group: 'Grupo Utilitario',
-        driver: 'Sergio Pérez',
-        numberOfTrailers: 1,
-        lastCodeSeal: '270070',
-        lastLocation: '25.752935, -100.5066483'
-    }, 
-    {
-        id :'16',
-        name: 'TRC226',
-        trailer : 'Remolque 17',
-        date: '19/03/2023 | 06:35:23',
-        status: 'En movimiento',
-        codeSeal : '270071',
-        group: 'Grupo Utilitario',
-        driver: 'Sergio Lopez',
-        numberOfTrailers: 2,
-        lastCodeSeal: '270071',
-        lastLocation: '25.752935, -100.5066483'
-    },
-
-];
-
-
-
-
-export const detailDevice =    {
-            id :'1',
-            name: 'PROTANKER2',
-            trailers : [
-                {   
-                    id: '1',
-                    name: 'Remolque 1',
-                    compartments : [
-                       {
-                        id: '12',
-                        name: 'Compartimiento 1',
-                        valve_box : 'open',
-                         wafer: 'close',
-                        dome: 'close',
-                        content: 'with_content'
-                       },
-                       {
-                        id: '13',
-                        name: 'Compartimiento 2',
-                        valve_box : 'close',
-                         wafer: 'close',
-                        dome: 'open',
-                        content: 'with_content'
-                       }
-                       
-                    ]
-
-                },],
-            data:[
-                {
-                    id: '1',
-                    date: '19/03/2023 | 06:35:23',
-                    codeSeal : '377852',
-                    numberCompartment:1,
-                    motorStatus: 'Detenido',
-                    capacityDevice: '30 000', 
-                    lastLocation: [25.752935, -100.5066483]
-                },
-                {
-                    id: '2',
-                    date: '19/03/2023 | 06:35:23',
-                    codeSeal : '270066',
-                    numberCompartment:1,
-                    motorStatus: 'Detenido',
-                    capacityDevice: '30 000', 
-                    lastLocation: [25.752935, -100.5066483]
-                },
-                {
-                    id:'3',
-                    date: '19/03/2023 | 06:35:23',
-                    codeSeal : '270066',
-                    numberCompartment:1,
-                    motorStatus: 'Detenido',
-                    capacityDevice: '30 000', 
-                    lastLocation: [25.752935, -100.5066483]
-
-                },
-                {
-                    id: '4',
-                    date: '19/03/2023 | 06:35:23',
-                    codeSeal : '270066',
-                    numberCompartment:1,
-                    motorStatus: 'Detenido',
-                    capacityDevice: '30 000', 
-                    lastLocation: [25.752935, -100.5066483]
-                },
-                {
-                    id: '5',
-                    date: '19/03/2023 | 06:35:23',
-                    codeSeal : '270066',
-                    numberCompartment:1,
-                    motorStatus: 'Detenido',
-                    capacityDevice: '30 000',
-                    lastLocation: [25.752935, -100.5066483]
-                },
-                {
-                    id: '6',
-                    date: '19/03/2023 | 06:35:23',
-                    codeSeal : '270066',
-                    numberCompartment:1,
-                    motorStatus: 'Detenido',
-                    capacityDevice: '30 000',
-                    lastLocation: [25.752935, -100.5066483]
-                },
-                {
-                    id: '7',
-                    date: '19/03/2023 | 06:35:23',
-                    codeSeal : '270066',
-                    numberCompartment:1,
-                    motorStatus: 'Detenido',
-                    capacityDevice: '30 000',
-                    lastLocation: [25.752935, -100.5066483]
-                }
-
-            ] 
-        
-    };
-export const devicesColumns=[
-    {
-        id:'unit',
-        field:'device',
-        header: 'Unidad'
-    }
-]
-export const detailDeviceColumns = [
-    {
-        id: '1',
-        field: 'date',
-        header: 'Fecha'
-    },
-    {
-        id: '2',
-        field: 'codeSeal',
-        header: 'Código de sello'
-    },
-    {
-        id: '3',
-        field: 'numberCompartment',
-        header: 'Compartimiento'
-    },
-    {
-        id: '4',
-        field: 'motorStatus',
-        header: 'Estado del motor'
-    },
-    {
-        id: '5',
-        field: 'capacityDevice',
-        header: 'Capacidad de la unidad'
-    },
-    {
-        id: '6',
-        field: 'lastLocation',
-        header: 'Últimas coordenadas'
-    }
-];
-
-export const columns = [
-    {
-        id: '1',
-        field: 'name',
-        header: 'Unidad'
-    },
-    {
-        id: '2',
-        field: 'trailer',
-        header: 'Remolque'
-    },
-    {
-        id: '3',
-        field: 'date',
-        header: 'Fecha'
-    },
-    {
-        id: '4',
-        field: 'status',
-        header: 'Estatus unidad'
-    },
-    {
-        id: '5',
-        field: 'codeSeal',
-        header: 'Código de sello'
-    }
-];
-
-export const detailedColumns = [
-    {
-        id: '1',
-        field:'name',
-        header: 'Unidad'
-    },
-    {
-        id: '2',
-        field: 'group',
-        header: 'Grupo'
-    },
-    {
-        id: '3',
-        field: 'driver',
-        header: 'Conductor'
-    },
-    {
-        id: '4',
-        field: 'numberOfTrailers',
-        header: 'Remolque'
-    },
-    {
-        id: '5',
-        field: 'lastCodeSeal',
-        header: 'Último sello'
-    },
-    {
-        id: '6',
-        field: 'lastLocation',
-        header: 'Última ubicación'
-    }
-];
-export const devicesAll = [
-    {  name: 'TRC213', code: 'dev1' },
-    {  name: 'TRC214', code: 'dev2' },
-    {  name: 'TRC215', code: 'dev3' },
-    {  name: 'TRC216', code: 'dev4' },
-    {  name: 'TRC217', code: 'dev5' },
-    {  name: 'TRC218', code: 'dev6' },
-    {  name: 'TRC219', code: 'dev7' },
-    {  name: 'TRC220', code: 'dev8' },
-    {  name: 'TRC221', code: 'dev9' },
-    {  name: 'TRC222', code: 'dev10' },
-    {  name: 'TRC217', code: 'dev5' }
-];
